Replace transaction action "Export" by 2 actions ("Export" and "Export (Transaction Tool 1.0)")
